pyproject.toml
[tool.poetry]
name = "cloud-storage"
version = "0.1.0"
description = ""
authors = ["ratmeow <rtm_bis@mail.ru>"]
readme = "README.md"
packages = [{include = "cloud_storage", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
bcrypt = "^5.0.0"
sqlalchemy = "^2.0.44"
pydantic = "^2.12.4"
python-dotenv = "^1.2.1"
asyncpg = "^0.30.0"
alembic = "^1.17.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
